dodany BubbleSort, trzeba zrobić suwak od prędkości jednej operacji(przesłać wartość z suwaka do algorytmów i wrzucić do chrono

+++ AlgorytmSortowania.h
@@ -16,7 +16,7 @@ protected:
     int operations = 0;
 public:
     virtual ~SortingAlgorithm() = default;
-    virtual void sort(std::vector<int>& data, std::mutex& mtx, bool& paused, bool& reset, MainWindow* window) = 0;
+    virtual void sort(std::vector<float>& data, std::mutex& mtx, bool& paused, bool& reset, MainWindow* window) = 0;
     int getOperations() const { return operations; }
 };
